requirements.txt
ollama
httpx
pydantic
python-dotenv
